EY-3207 Sjekke om bruker faktisk er attestant (AD groups)
