+++ apps/etterlatte-saksbehandling-ui/client/src/shared/api/user.ts
@@ -1,6 +1,6 @@
-import { ISaksbehandler } from '~store/reducers/InnloggetSaksbehandlerReducer'
+import { ISaksbehandler } from '~shared/types/saksbehandler'
 import { apiClient, ApiResponse } from './apiClient'
 
 export const hentInnloggetSaksbehandler = async (): Promise<ApiResponse<ISaksbehandler>> => {
-  return apiClient.get('/modiacontextholder/decorator')
+  return apiClient.get('/innlogget')
 }
