+++ apps/etterlatte-saksbehandling-ui/client/src/shared/types/saksbehandler.ts
@@ -1,0 +1,13 @@
+export interface ISaksbehandler {
+  ident: string
+  navn: string
+  fornavn: string
+  etternavn: string
+  enheter: IEnhet[]
+  kanAttestere: boolean
+}
+
+interface IEnhet {
+  enhetId: string
+  navn: string
+}
