+++ apps/etterlatte-saksbehandling-ui/server/src/utils/parsejwt.ts
@@ -1,6 +1,7 @@
 export interface Token {
   name: string
   NAVident: string
+  groups: string[]
   aud: string
   iss: string
   exp: number
@@ -8,7 +9,8 @@ export interface Token {
 }
 
 export const parseJwt = (token: string): Token => {
-  var base64Payload = token.split('.')[1]
-  var payload = Buffer.from(base64Payload, 'base64')
+  const base64Payload = token.split('.')[1]
+  const payload = Buffer.from(base64Payload, 'base64')
+
   return JSON.parse(payload.toString())
 }
